+++ src/main/java/com/caoccao/javet/interop/converters/JavetProxyConverter.java
@@ -64,9 +64,12 @@ public class JavetProxyConverter extends JavetObjectConverter {
         if (v8Value != null && !(v8Value.isUndefined())) {
             return (T) v8Value;
         }
-        boolean staticMode = config.isStaticClassEnabled() && object instanceof Class;
+        boolean classMode = false;
+        if (object instanceof Class) {
+            classMode = JavetUniversalProxyHandler.isClassMode((Class) object);
+        }
         V8ValueProxy v8ValueProxy;
-        if (staticMode) {
+        if (classMode) {
             try (V8ValueFunction v8ValueFunction = v8Runtime.createV8ValueFunction(DUMMY_FUNCTION_STRING)) {
                 v8ValueProxy = v8Runtime.createV8ValueProxy(v8ValueFunction);
             }
@@ -74,13 +77,8 @@ public class JavetProxyConverter extends JavetObjectConverter {
             v8ValueProxy = v8Runtime.createV8ValueProxy();
         }
         try (IV8ValueObject iV8ValueObjectHandler = v8ValueProxy.getHandler()) {
-            JavetUniversalProxyHandler<Object> javetUniversalProxyHandler;
-            if (staticMode) {
-                javetUniversalProxyHandler = new JavetUniversalProxyHandler<>(
-                        v8Runtime, null, (Class<Object>) object, true);
-            } else {
-                javetUniversalProxyHandler = new JavetUniversalProxyHandler<>(v8Runtime, object);
-            }
+            JavetUniversalProxyHandler<Object> javetUniversalProxyHandler =
+                    new JavetUniversalProxyHandler<>(v8Runtime, object);
             List<JavetCallbackContext> javetCallbackContexts = iV8ValueObjectHandler.bind(javetUniversalProxyHandler);
             iV8ValueObjectHandler.set(PROXY_TARGET, javetCallbackContexts.get(0).getHandle());
         }